RGB LED Controler powered by Fadecandy
To control RGB LED's powered by a Fadecandy controler.  Original purpose
is to control LED's inside a computer case uing different variables and
option.  An option will be to have the LED's change color depending on
the tempeture data from the GPU and/or CPU.  Another option would be to
just have a static Color if one would like, or to replicate the outside
boarder of a computer screen.